entrylens-api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic-settings==2.10.1
sqlalchemy==2.0.43
aiosqlite==0.21.0
python-multipart==0.0.20
supabase==2.10.0
pillow==11.1.0
numpy==2.2.3
insightface==0.7.3
onnxruntime==1.17.1
opencv-python-headless==4.9.0.80
